♻️ V18P6W acr: decompose command specs
