+++ .agentplane/tasks/202605290402-V18P6W/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605290402-V18P6W`
+Title: ACR command specs decomposition
+Canonical task record: `.agentplane/tasks/202605290402-V18P6W/README.md`
+
+## Summary
+
+ACR command specs decomposition
+
+Decompose packages/agentplane/src/commands/acr/acr.command.ts by extracting ACR command specs and parsed types into a focused acr.specs module, preserving public exports and CLI parsing behavior while reducing runtime hotspot count.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/commands/acr/acr.command.ts by extracting ACR command specs and parsed types into a focused acr.specs module, preserving public exports and CLI parsing behavior while reducing runtime hotspot count.
+- Out of scope: unrelated refactors not required for "ACR command specs decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T04:02:50.740Z
+- Branch: task/202605290402-V18P6W/acr-command-specs-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
