+++ .agentplane/tasks/202605290402-V18P6W/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 V18P6W task: ACR command specs decomposition [202605290402-V18P6W]
