+++ .agentplane/tasks/202605290402-V18P6W/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-29T04:02:50.740Z
+
+## Task
+
+- Task: `202605290402-V18P6W`
+- Title: ACR command specs decomposition
+- Status: DOING
+- Branch: `task/202605290402-V18P6W/acr-command-specs-decomposition`
+- Canonical task record: `.agentplane/tasks/202605290402-V18P6W/README.md`
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T04:02:50.740Z
+- Branch: task/202605290402-V18P6W/acr-command-specs-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
